+++ mvpipe/runner/slurm.py
@@ -176,8 +176,9 @@ class SlurmRunner(Runner):
         src += '    RETVAL=$?\n'
         src += '    if [ $RETVAL -ne 0 ]; then\n'
         for out in job.outputs:
-            if out[0] != '.':
-                src += '        if [ -e "%s" ]; then rm "%s"; fi\n' % (out, out)
+            if not 'keepfailed' in jobopts or not jobopts['keepfailed']:
+                if out[0] != '.':
+                    src += '        if [ -e "%s" ]; then rm "%s"; fi\n' % (out, out)
         src += '    fi\n'
         src += '    exit $RETVAL\n'
         src += 'fi\n'
